pyproject.toml
[project]
name = "unemployment-speedrun"
version = "0.1.0"
description = "AI job application agent"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.44.0",
    "gspread>=6.1.0",
    "google-auth>=2.29.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "anthropic>=0.28.0",
    "streamlit>=1.35.0",
    "reportlab>=4.2.0",
    "keyring>=25.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
